implement document api in Thing::Api::Create, etc. not sure if this will be final.
