+++ config/routes.rb
@@ -26,5 +26,6 @@ Rails.application.routes.draw do
 
   namespace :api do
     resources :things
+    resources :users
   end
 end
