+++ test/api/things_test.rb
@@ -10,7 +10,7 @@ class ApiThingsTest < MiniTest::Spec
 
   it do
     thing = Thing::Create.(thing: {name: "Rails"}).model
-    get "/api/things/#{thing.id}", "CONTENT_TYPE" => "application/json"
+    get "/api/things/#{thing.id}", "", "CONTENT_TYPE" => "application/json", "HTTP_ACCEPT"=>"application/json"
     last_response.body.must_equal %{{"id":#{thing.id},"name":"Rails","links":[{"rel":"self","href":"/things/#{thing.id}"}],"authors":[],"comments":[]}}
   end
 
@@ -18,10 +18,11 @@ class ApiThingsTest < MiniTest::Spec
     data = {name: "Lotus"}
     # request.env["HTTP_ACCEPT"] = "application/json"
     post "/api/things/", data.to_json, "CONTENT_TYPE" => "application/json", "HTTP_ACCEPT"=>"application/json"
+    id = Thing.last.id
 
-    last_response.headers["Location"].must_equal "http://example.org/api/things/1"
+    last_response.headers["Location"].must_equal "http://example.org/api/things/#{id}"
     assert last_response.created?
-    last_response.body.must_equal %{{"name":"Lotus","authors":[],"id":1,"links":[{"rel":"self","href":"/api/things/1"}]}}
+    last_response.body.must_equal %{{"name":"Lotus","authors":[],"id":#{id},"links":[{"rel":"self","href":"/api/things/#{id}"}]}}
 
     # assert_equal "http://example.org/redirected", last_request.url
     # assert last_response.ok?
